games

+++ src/api/index.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const API = axios.create({
-    baseURL: 'http://34.124.217.226:7000/admin',
+    baseURL: 'http://34.124.217.226:7000',
     //   withCredentials: true,
     // credentials: 'include'
 });
@@ -47,27 +47,38 @@ export const apiInstance = API;
 
 export const BackEndAddress = 'http://34.124.217.226:7000'
 // ----------------------------User (staff) --------------------------
-export const getAllUsers = () => API.get('/user');
-export const getUser = (identity) => API.get(`/user/${identity}`);
-export const createUser = (newUser) => API.post('/user', newUser);
-export const updateUser = (identity, updatedUser) => API.put(`/user/${identity}`, updatedUser);
-export const deleteUser = (identity) => API.delete(`/user/${identity}`);
+export const getAllUsers = () => API.get('/admin/user');
+export const getUser = (identity) => API.get(`/admin/user/${identity}`);
+export const createUser = (newUser) => API.post('/admin/user', newUser);
+export const updateUser = (identity, updatedUser) => API.put(`/admin/user/${identity}`, updatedUser);
+export const deleteUser = (identity) => API.delete(`/admin/user/${identity}`);
 // ----------------------------User (staff) --------------------------
-export const getAllGames = () => API.get('/game');
-export const getGame = (identity) => API.get(`/game/${identity}`);
-export const createGame = (newGame) => API.post('/game', newGame);
-export const updateGame = (identity, updatedGame) => API.put(`/game/${identity}`, updatedGame);
-export const deleteGame = (identity) => API.delete(`/game/${identity}`);
+export const getAllGames = () => API.get('/admin/game');
+export const getGame = (identity) => API.get(`/admin/game/${identity}`);
+export const createGame = (newGame) => API.post('/admin/game', newGame);
+export const updateGame = (identity, updatedGame) => API.put(`/admin/game/${identity}`, updatedGame);
+export const deleteGame = (identity) => API.delete(`/admin/game/${identity}`);
 // ---------------------------- Auth --------------------------
 // export const getAllUsers = () => API.get('/login');
 // ---------------------------- Statistics --------------------------
-export const getCountAllUsers = () => API.get('/statistic/total_users');
-export const getCountAllGames = () => API.get('/statistic/total_games');
-export const getCountAllEnterprises = () => API.get('/statistic/total_enterprises');
-export const getCountAllEndUsers = () => API.get('/statistic/total_end_users');
+export const getCountAllUsers = () => API.get('/admin/statistic/total_users');
+export const getCountAllGames = () => API.get('/admin/statistic/total_games');
+export const getCountAllEnterprises = () => API.get('/admin/statistic/total_enterprises');
+export const getCountAllEndUsers = () => API.get('/admin/statistic/total_end_users');
+export const getStatisticUsers = () => API.get('/admin/statistic/total_new_end_users_in_week')
+export const getStatisticEnterprises = () => API.get('/admin/statistic/total_new_enterprises_in_week')
+
 
 // ----------------------------image --------------------------
-export const uploadGameImage = () => API.get('/image/game');
 export const getGameImage = (identity) => API.get(`/image/game/${identity}`);
 
+export const uploadGameImage = (file) => {
+    const formData = new FormData();
+    formData.append('image', file);
 
+    return API.post('/image/game/upload', formData, {
+        headers: {
+            'Content-Type': 'multipart/form-data',
+        },
+    });
+};
